+++ Sis_Vendas_Mega/Model/Score.cs
@@ -10,10 +10,10 @@ namespace Sis_Vendas_Mega.Model
         public DateTime? DepartureTime { get; private set; }
         public TimeSpan? Worked { get; private set; }
         public int EmployeeId { get; private set; }
-        public int Code { get; private set; }
+        public long Code { get; private set; }
         public virtual Employee Employee { get; private set; }
 
-        public Score(int employeeId, int code, DateTime? entryTime)
+        public Score(int employeeId, long code, DateTime? entryTime)
         {
             EmployeeId = employeeId;
             Code = code;
